46 refactor the code as it is growing larger (#89)
* Refactoring

* Removed unncessary files
