+++ include/build_master/meson_build_gen.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string_view>
+
+void RegenerateMesonBuildScript(std::string_view directory = "", bool isForce = false);
