+++ include/build_master/pre_config_script.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string_view>
+
+bool RunPreConfigScript(std::string_view directory);
