+++ include/build_master/version.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+// NOTE: Following defines are Automatically Defined by the Build System (meson.build)
+// BUILDMASTER_VERSION_MAJOR
+// BUILDMASTER_VERSION_MINOR
+// BUILDMASTER_VERSION_MICRO
+// MESON_BUILD_TEMPLATE_PATH
+
+#define BUILDMASTER_VERSION_ENCODE(major, minor, micro) (((major) * 10000) + ((minor) * 100) + ((micro) * 1))
+#define BUILDMASTER_VERSION BUILDMASTER_VERSION_ENCODE(BUILDMASTER_VERSION_MAJOR, BUILDMASTER_VERSION_MINOR, BUILDMASTER_VERSION_MICRO)
+
+#define BUILDMASTER_VERSION_XSTRINGIZE(major, minor, micro) #major"."#minor"."#micro
+#define BUILDMASTER_VERSION_STRINGIZE(major, minor, micro) BUILDMASTER_VERSION_XSTRINGIZE(major, minor, micro)
+#define BUILDMASTER_VERSION_STRING BUILDMASTER_VERSION_STRINGIZE(BUILDMASTER_VERSION_MAJOR, BUILDMASTER_VERSION_MINOR, BUILDMASTER_VERSION_MICRO)
