+++ source/meson_build_gen.cpp
@@ -1,3 +1,7 @@
+#include <build_master/misc.hpp> // for LoadTextFile(), and GetPathStrRelativeToDir()
+#include <build_master/json_parse.hpp> // for ParseBuildMasterJson(), and GetJsonKeyValue<>()
+#include <build_master/version.hpp>
+
 #include <iostream>
 #include <cstdlib>
 #include <format>
@@ -6,49 +10,11 @@
 #include <string>
 #include <string_view>
 #include <type_traits>
-#include <sstream>
+#include <fstream>
 
-#include <CLI/CLI.hpp>
-#include <nlohmann/json.hpp>
 #include <spdlog/spdlog.h>
-#include <invoke/invoke.hpp>
-#include <common/defines.hpp> // for com::to_upper()
 
-// NOTE: Following defines are Automatically Defined by the Build System (meson.build)
-// BUILDMASTER_VERSION_MAJOR
-// BUILDMASTER_VERSION_MINOR
-// BUILDMASTER_VERSION_MICRO
-// MESON_BUILD_TEMPLATE_PATH
-
-#define BUILDMASTER_VERSION_ENCODE(major, minor, micro) (((major) * 10000) + ((minor) * 100) + ((micro) * 1))
-#define BUILDMASTER_VERSION BUILDMASTER_VERSION_ENCODE(BUILDMASTER_VERSION_MAJOR, BUILDMASTER_VERSION_MINOR, BUILDMASTER_VERSION_MICRO)
-
-#define BUILDMASTER_VERSION_XSTRINGIZE(major, minor, micro) #major"."#minor"."#micro
-#define BUILDMASTER_VERSION_STRINGIZE(major, minor, micro) BUILDMASTER_VERSION_XSTRINGIZE(major, minor, micro)
-#define BUILDMASTER_VERSION_STRING BUILDMASTER_VERSION_STRINGIZE(BUILDMASTER_VERSION_MAJOR, BUILDMASTER_VERSION_MINOR, BUILDMASTER_VERSION_MICRO)
-
-using json = nlohmann::ordered_json; 
-
-static constexpr std::string_view gBuildMasterJsonFilePath = "build_master.json";
 static constexpr std::string_view gMesonBuildScriptFilePath = "meson.build";
-static constexpr std::string_view gBash = "bash";
-static constexpr std::string_view gMesonExecutableName = "build_master_meson";
-
-
-// directoryBase: The base directory against which the the final path need to be calculated
-// relativeDir: The path relative to the base directory passed as 'directoryBase' (first arg)
-static std::string GetPathStrRelativeToDir(std::string_view directoryBase, std::string_view relativePath)
-{
-	auto buildMasterJsonFilePath = std::filesystem::path(directoryBase) / std::filesystem::path(relativePath);
-	return buildMasterJsonFilePath.string();
-}
-
-// Use this function whenever you meant to get gBuildMasterJsonFilePath.
-// DO NOT use gBuildMasterJsonFilePath directly as that would not consider the --directory flag 
-static std::string GetBuildMasterJsonFilePath(std::string_view directory = ".")
-{
-	return GetPathStrRelativeToDir(directory, gBuildMasterJsonFilePath);
-}
 
 // Use this function whenever you meant to get gMesonBuildScriptFilePath.
 // DO NOT use gMesonBuildScriptFilePath directory as that would not consider the --directory flag 
@@ -67,73 +33,6 @@ static bool IsRegenerateMesonBuildScript(std::string_view directory)
 		return buildMasterJsonTime >= mesonBuildScriptTime;
 	}
  	return true;
-}
-
-static std::string LoadTextFile(std::string_view filePath)
-{
-	std::ifstream stream(filePath.data());
-	if(!stream.is_open())
-	{
-		std::cerr << "Error: Failed to open " << filePath << "\n";
-		exit(EXIT_FAILURE);
-	}
-    std::string line;
-    std::string contents;
-    while (std::getline(stream, line)) {
-        contents.append(line);
-        contents.append("\n");
-    }
-    return contents;	
-}
-
-static std::optional<std::pair<std::size_t, std::size_t>> IsBlankLine(const std::string& str, std::size_t charIndex)
-{
-	std::size_t lineBegin;
-	while(charIndex > 0)
-	{
-		char ch = str[--charIndex];
-		if(ch == '\n')
-		{
-			lineBegin = charIndex;
-			break;
-		}
-		if(!std::isspace(ch))
-			return { };
-	}
-	std::size_t lineEnd;
-	while(charIndex < str.size())
-	{
-		char ch = str[charIndex++];
-		if(ch == '\n')
-		{
-			lineEnd = charIndex;
-			break;
-		}
-		if(!std::isspace(ch))
-			return { };
-	}
-	return { { lineBegin, lineEnd } };
-}
-
-static void EraseCppComments(std::string& str)
-{
-	std::size_t index;
-	while((index = str.find("//")) != std::string::npos)
-	{
-		auto newLineIndex = str.find_first_of('\n', index + 2);
-		str.erase(std::next(str.begin(), index), std::next(str.begin(), newLineIndex));
-		if(auto result = IsBlankLine(str, index); result.has_value())
-			str.erase(std::next(str.begin(), result->first), std::next(str.begin(), result->second));
-	}
-}
-
-// directory: value passed to --directory flag
-static json ParseBuildMasterJson(std::string_view directory)
-{
-	std::string jsonStr = LoadTextFile(GetBuildMasterJsonFilePath(directory));
-	EraseCppComments(jsonStr);
-	json data = json::parse(jsonStr);
-	return data;
 }
 
 static std::string LoadMesonBuildScriptTemplate()
@@ -264,24 +163,6 @@ static TargetType DetectTargetType(const json& targetJson)
 	if(it != targetJson.end() && it.value().template get<bool>())
 		return TargetType::HeaderOnlyLibrary;
 	return TargetType::Executable;
-}
-
-template<typename T>
-static std::optional<T> GetJsonKeyValueOrNull(const json& jsonObj, std::string_view key)
-{
-	if(auto it = jsonObj.find(key); it != jsonObj.end())
-		return { it.value().template get<T>() };
-	return { };
-}
-
-template<typename T>
-static T GetJsonKeyValue(const json& jsonObj, std::string_view key, std::optional<T> defaultValue = {})
-{
-	if(auto result = GetJsonKeyValueOrNull<T>(jsonObj, key); result.has_value())
-		return result.value();
-	if(!defaultValue.has_value())
-		throw std::runtime_error(std::format("No such key: {} exists in the json provided, and default value is not given either", key));
-	return std::move(defaultValue.value());
 }
 
 static void ProcessStringList(const json& jsonObj, std::ostringstream& stream, std::string_view delimit = " ", std::optional<TokenTransformCallback> callback = { })
@@ -582,7 +463,7 @@ static void GenerateMesonBuildScript(std::string_view directory)
 }
 
 // directory: value passed to --directory flag
-static void RegenerateMesonBuildScript(std::string_view directory = "", bool isForce = false)
+void RegenerateMesonBuildScript(std::string_view directory, bool isForce)
 {
 	if(!std::filesystem::exists(GetBuildMasterJsonFilePath(directory)))
 	{
@@ -595,334 +476,4 @@ static void RegenerateMesonBuildScript(std::string_view directory = "", bool isF
 		GenerateMesonBuildScript(directory);
 	else
 		std::cout << "Info: meson.build is upto date\n";
-}
-
-// Stores values of the arguments passed to 'init' command
-// Example: build_master init --name=BufferLib --canonical_name=bufferlib --force
-struct ProjectInitCommandArgs
-{
-	// --name=<your project name>
-	// NOTE: This can't be empty
-	std::string_view projectName;
-	// --canonical_name=<
-	// NOTE: This can't be empty
-	std::string_view canonicalName;
-	// --directory=/path/to/a/directory
-	// NOTE: This can be empty
-	std::string_view directory;
-	// --create-cpp
-	bool isCreateCpp { false };
-	// --force
-	bool isForce { false };
-};
-
-// Create a new text file and writes the text data passed as argument
-static void WriteTextFile(const std::string_view filePath, const std::string_view textData)
-{
-	std::ofstream file(filePath.data(), std::ios_base::trunc);
-	if(!file.is_open())
-	{
-		std::cout << std::format("Error: Failed to create {} file\n", filePath);
-		exit(EXIT_FAILURE);
-	}
-	file << textData << std::endl;
-	file.close();
-}
-
-// Creates a text file with initData if doesn't exist already. It also creates any intermediate directories if doesn't exist.
-static void CreateFile(const std::filesystem::path& path, std::string_view initData)
-{
-	auto pathCopy = path;
-	auto dirs = pathCopy.remove_filename();
-	try
-	{
-		std::filesystem::create_directories(dirs);
-	} catch(const std::exception& except)
-	{
-		std::cerr << "Error: Failed to create " << dirs.string() << ", " << except.what();
-	}
-	if(std::filesystem::exists(path))
-		std::cout << std::format("Info: file \"{}\" already exists, skipping creating\n", path.string());
-	else
-		WriteTextFile(path.string(), initData);
-}
-
-using namespace std::literals;
-
-// include/<canonical_name>/api_defines.h or .hpp
-// include/<canonical_name>/defines.h or .hpp
-// source/main.cpp
-static void GenerateSeedFiles(const ProjectInitCommandArgs& args)
-{
-	constexpr std::string_view mainSourceInitData[2] = {
-R"(
-#include <iostream>
-
-int main()
-{
-	std::cout << "Hello World";
-	return 0;
-}
-)",
-R"(
-#include <stdio.h>
-
-int main()
-{
-	puts("Hello World");
-	return 0;
-}
-)" };
-	constexpr std::string_view apiDefinesInitTemplate =
-R"(
-#pragma once
-
-#if (defined _WIN32 || defined __CYGWIN__) && defined(__GNUC__)
-#	define {0}_IMPORT_API __declspec(dllimport)
-#	define {0}_EXPORT_API __declspec(dllexport)
-#else
-#	define {0}_IMPORT_API __attribute__((visibility("default")))
-#	define {0}_EXPORT_API __attribute__((visibility("default")))
-#endif
-
-#ifdef {0}_BUILD_STATIC_LIBRARY
-#	define {0}_API
-#elif defined({0}_BUILD_DYNAMIC_LIBRARY)
-#	define {0}_API {0}_EXPORT_API
-#elif defined({0}_USE_DYNAMIC_LIBRARY)
-#	define {0}_API {0}_IMPORT_API
-#elif defined({0}_USE_STATIC_LIBRARY)
-#	define {0}_API
-#else
-#	define {0}_API
-#endif
-)";
-	constexpr std::string_view definesInitTemplate =
-R"(
-#pragma once
-#include <{0}/api_defines.{1}>
-#if !defined({2}_RELEASE) && !defined({2}_DEBUG)
-#   warning "None of {2}_RELEASE && {2}_DEBUG is defined; using {2}_DEBUG"
-#   define {2}_DEBUG
-#endif
-)";
-	auto canonicalNameUpper = com::to_upper(args.canonicalName);
-	std::string apiDefinesInitData = std::format(apiDefinesInitTemplate, canonicalNameUpper);
-	std::string_view headerFileExt { args.isCreateCpp ? "hpp" : "h" };
-	std::string definesInitData = std::format(definesInitTemplate, args.canonicalName, headerFileExt, canonicalNameUpper);
-	auto sourceDir = std::filesystem::path(args.directory) / "source"sv;
-	auto includeDir = std::filesystem::path(args.directory) / "include"sv / args.canonicalName;
-	if(args.isCreateCpp)
-	{
-		CreateFile(sourceDir / "main.cpp", mainSourceInitData[0]);
-		CreateFile(includeDir / "api_defines.hpp", apiDefinesInitData);
-		CreateFile(includeDir / "defines.hpp", definesInitData);
-	}
-	else
-	{
-		CreateFile(sourceDir / "main.c", mainSourceInitData[1]);
-		CreateFile(includeDir / "api_defines.h", apiDefinesInitData);
-		CreateFile(includeDir / "defines.h", definesInitData);
-	}
-}
-
-//  build_master init
-static void IntializeProject(const ProjectInitCommandArgs& args)
-{
-	std::cout << std::format("Project Name = \"{}\", Canonical Name = \"{}\"", args.projectName, args.canonicalName) << "\n";
-	const auto filePathStr =  GetBuildMasterJsonFilePath(args.directory);
-	if(!args.isForce && std::filesystem::exists(filePathStr))
-	{
-		std::cerr << "Error: build_master.json already exists, please remove it first or pass --force to overwrite it\n";
-		exit(EXIT_FAILURE);
-	}
-	json buildMasterJson = 
-	{
-		{ "project_name" , args.projectName },
-		{ "canonical_name", args.canonicalName },
-		{ "include_dirs", "include" },
-		{ "targets",
-			{			
-				{ 
-					{ "name", args.canonicalName },
-					{ "is_executable", true },
-					{ "sources", { args.isCreateCpp ? "source/main.cpp" : "source/main.c" } }
-				}
-			}
-		}
-	};
-	WriteTextFile(filePathStr, buildMasterJson.dump(4, ' ', true));
-	std::cout << std::format("Success: {} is generated\n", filePathStr);
-
-	GenerateSeedFiles(args);
-}
-
-static std::ostream& operator<<(std::ostream& stream, const std::vector<std::string>& v)
-{
-	for(const auto& value : v)
-		stream << value << " ";
-	return stream;
-}
-
-static bool RunPreConfigScript(std::string_view directory)
-{
-	json buildMasterJson = ParseBuildMasterJson(directory);
-	if(auto result = GetJsonKeyValueOrNull<std::string>(buildMasterJson, "pre_config_hook"); result.has_value())
-	{
-		spdlog::info("Running pre-config hook script");
-		auto returnCode = invoke::Exec({ std::string { gBash }, result.value() }, directory);
-		if(returnCode < 0)
-		{			
-			spdlog::error("pre-config hook script returned non-zero code");
-			exit(returnCode);
-		}
-		spdlog::info("pre-config hook script run success");
-		return true;
-	}
-	return false;
-}
-
-static std::string SelectPath(const std::vector<std::string>& paths)
-{
-	#ifdef _WIN32
-		// First try mingw
-		for(const auto& path : paths)
-			if(path.find("mingw") != std::string::npos)
-				return path;
-		// Then fallback to msys
-		for(const auto& path : paths)
-			if(path.find("msys") != std::string::npos)
-				return path;
-		spdlog::error("Couldn't find mingw or msys equivalent executable path");
-		exit(EXIT_FAILURE);
-	#else
-		return paths[0];
-	#endif
-}
-
-// cmdName: name of the command (not the full path, just the name)
-// restArgs: The rest of the arguments which need to be passed to that cmd
-static std::vector<std::string> BuildArgumentsForCmdRun(std::string_view cmdName, const std::vector<std::string>& restArgs)
-{
-  	std::vector<std::string> finalArgs;
-  	finalArgs.reserve(restArgs.size() + 1);
-  	auto paths = invoke::GetExecutablePaths(cmdName);
-  	if(!paths)
-  	{
-  		spdlog::error("Couldn't find paths for the executable: {}", cmdName);
-  		exit(EXIT_FAILURE);
-  	}
-  	auto fullMesonExePath = SelectPath(paths.value());;
-  	finalArgs.push_back(fullMesonExePath);
-  	for(const auto& arg : restArgs)
-  		finalArgs.push_back(arg);
-  	return finalArgs;
-}
-
-// cmdName: Just the name of the command
-// restArgs: The rest of the argumentst which need to be passed to that cmd
-static decltype(auto) RunCmd(std::string_view cmdName, std::string_view workDirectory, const std::vector<std::string>& restArgs)
-{
-  	// Prepare arguments
-  	std::vector<std::string> finalArgs = BuildArgumentsForCmdRun(cmdName, restArgs);
-  	// Logging
-  	std::stringstream strstream;
-  	strstream << "Command: " << finalArgs << "\n";
-  	spdlog::info(strstream.str());
-  	return invoke::Exec(finalArgs, workDirectory);	
-}
-
-// build_master meson
-// directory: value passed to --directory flag
-static void InvokeMeson(std::string_view directory, const std::vector<std::string>& args)
-{
-	// Ensure the meson.build script is upto date
-	RegenerateMesonBuildScript(directory);
-	// Run pre-configure script if 'meson setup' command is executed
-	if(args.size() == 0 || args[0] != "setup")
-		RunPreConfigScript(directory);
-	// Run the meson command
-  	exit(RunCmd(gMesonExecutableName, directory, args));
-}
-
-static void PrintVersionInfo() noexcept
-{
-	std::cout << "Build Master " << BUILDMASTER_VERSION_STRING << "\n";
-	#ifdef BUILD_MASTER_RELEASE
-		std::cout << "Build Type: Release\n";
-	#else
-		std::cout << "Build Type: Debug\n";
-	#endif
-}
-
-int main(int argc, const char* argv[])
-{
-	CLI::App app;
-
-	bool isPrintVersion = false, 
-		isUpdateMesonBuild = false,
-		isMesonBuildTemplatePath = false,
-		isExecutePreConfigHook = false,
-		isForce = false;
-	std::string directory;
-	app.add_flag("--version", isPrintVersion, "Prints version number of Build Master");
-	app.add_flag("--update-meson-build", isUpdateMesonBuild, "Regenerates the meson.build script if the build_master.json file is more recent");
-	app.add_flag("--meson-build-template-path", isMesonBuildTemplatePath, "Prints lookup path of meson.build.template, typically this is only for debugging purpose");
-	app.add_flag("--execute-pre-config-hook", isExecutePreConfigHook, "Executes pre_config_hook (shell script) if any");
-	app.add_flag("--force", isForce, "if --update-meson-build flag is present along with this --force then meson.build script is generated even if it is upto date");
-	app.add_option("--directory", directory, "Directory path in which to look for build_master.json, by default it is the current working directory");
-
-	// Project Initialization Sub command	
-	{
-		CLI::App* scInit = app.add_subcommand("init", "Project initialization, creates build_master.json file");
-		ProjectInitCommandArgs args;
-		scInit->add_option("--name", args.projectName, "Name of the Project")->required();
-		scInit->add_option("--canonical_name", args.canonicalName, 
-			"Conanical name of the Project, typically used for naming files")->required();
-		scInit->add_flag("--force", args.isForce, "Forcefully overwrites the existing build_master.json file, use it carefully");
-		scInit->add_flag("--create-cpp", args.isCreateCpp, "Create source/main.cpp instead of source/main.c, use this flag if you intend to write C++ code in start");
-		scInit->add_option("--directory", args.directory, "Directory path in which build_master.json and source/main.cpp would be created, by default it is the current working directory");
-		scInit->callback([&]()
-		{
-			// If --directory option for this `init` command is not supplied while it is supplied for build_master
-			// then use the build_master's --directory option's value.
-			// We considering the following case here:
-			//	build_master --directory init --name=myProject --canonical_name=myproject
-			if(args.directory.size() == 0)
-				args.directory = directory; 
-			IntializeProject(args); 
-		});
-	}
-	
-	// Meson Sub command
-	{
-		CLI::App* scMeson = app.add_subcommand("meson", "Invokes meson build system (cli), all the arguments passed to this subcommands goes to actual meson command");
-		scMeson->allow_extras();
-		scMeson->callback([&]() { InvokeMeson(directory, scMeson->remaining()); });
-	}
-
-	CLI11_PARSE(app, argc, argv);
-	
-	if(isPrintVersion)
-	{
-		PrintVersionInfo();
-		return EXIT_SUCCESS;
-	}
-	if(isUpdateMesonBuild)
-	{
-		RegenerateMesonBuildScript(directory, isForce);
-		return EXIT_SUCCESS;
-	}
-	if(isMesonBuildTemplatePath)
-	{
-		spdlog::info("MESON_BUILD_TEMPLATE_PATH: {}", MESON_BUILD_TEMPLATE_PATH);
-		return EXIT_SUCCESS;
-	}
-	if(isExecutePreConfigHook)
-	{
-		if(!RunPreConfigScript(directory))
-			spdlog::info("No pre_config_hook to run");
-	}
-
-	return EXIT_SUCCESS;
 }
